caxselling/pca/src/requirements.txt
pydantic == 2.8.2
pyYAML == 6.0.2
flask-restx == 1.3.0
flask == 2.2.5
Flask-Cors == 5.0.0
swagger-ui-bundle == 0.0.5
python_dateutil == 2.8.2
setuptools >= 75.8.0
paho-mqtt == 1.5.1
kafka-python == 2.0.2
jinja2
cython
numpy
scikit-learn
scikit-learn-intelex == 2024.7.0
daal-devel == 2024.7.0
daal-include == 2024.7.0
daal-static == 2024.7.0
daal4py == 2024.7.0
modin[all] == 0.32.0
psycopg2-binary == 2.9.10
